+++ conf.json
@@ -1,0 +1,6 @@
+{
+    "source": {
+      "include": ["."],
+      "exclude": ["node_modules/"]
+    }
+  }
