+++ index.html
@@ -16,7 +16,7 @@
                 <a href="index.html" title="Home"><img src="images/PP_logo_v2.png" class="logo"></a>
             </div>
             <nav>
-                <a href="sign-up-popup" class="nav-item">Tour</a>
+                <!-- <a href="sign-up-popup" class="nav-item">Tour</a> -->
                 <a href="signup.html" class="nav-item">Sign Up</a>
                 <a href="login.html" class="nav-item">Log In</a>
             </nav>
@@ -29,7 +29,7 @@
                     Our purpose is to share the most essential and insightful knowledge with one another.
                     The community language is English, and the platform is free to use, forever.
                     We welcome all friendly product people, and we reward the best contributors.</p>
-                <a href="signup.html" title="Sign Up Page" class="cta">Join our geeky product world!</a>
+                <a href="signup.html" title="Sign Up Page" class="cta" id="join-geeky-cta">Join our geeky product world!</a>
             </section>
             <section class="product-passion-concept-section">
                 <div class="intro-box">
