+++ js/constants/localStorage.mjs
@@ -1,0 +1,5 @@
+const usernameConst = localStorage.getItem("username");
+const token = localStorage.getItem("accessToken");
+const profilePictureUrl = localStorage.getItem("profilePictureUrl");
+
+export { usernameConst, token, profilePictureUrl };
