+++ js/login.mjs
@@ -1,3 +1,5 @@
+import { baseUrl } from "./constants/api.mjs";
+
 // Login validation
 
 const form = document.querySelector(".login-form");
@@ -13,16 +15,15 @@ const passwordCont = document.querySelector(".password-feedback-container");
 const messageCont = document.querySelector(".message-cont");
 const submitButton = document.querySelector("#submit-button");
 
-const baseUrl = "https://api.noroff.dev/api/v1";
 const loginUrlEndPoint = "/social/auth/login";
-const loginUrl = baseUrl + loginUrlEndPoint;
+const loginUrl = `${baseUrl}/social/auth/login`;
 
 /* Log In Code */
 
 form.addEventListener("submit", async (data) => {
   data.preventDefault();
 
-  localStorage. clear();
+  localStorage.clear();
 
   if (!passwordInput.value || !emailInput.value) {
     messageCont.innerHTML = `<span class="error-message">Please fill in your email and password to log in.</span>`;
@@ -50,15 +51,18 @@ form.addEventListener("submit", async (data) => {
         
     try {
       const resp = await fetch(loginUrl, requestOptions);
-      console.log(resp);
-
       const json = await resp.json();
 
-      console.log("Login API reponse: ", json);
+      if (!resp.ok) {
+				throw new Error(json.errors[0].message);
+			}
 
       const token = json.accessToken;
-      console.log("AccessToken const: ", token);
+      const profilePictureUrl = json.avatar;
+      const username = json.name;
+      localStorage.setItem("username", username);
       localStorage.setItem("accessToken", token);
+      localStorage.setItem("profilePictureUrl", profilePictureUrl);
 
       if (resp.ok) {
         const profileOptions = {
@@ -67,24 +71,13 @@ form.addEventListener("submit", async (data) => {
           },
         }
 
-        const username = json.name;
-
-        localStorage.setItem("username", username);
-
-        console.log("Username: ", username)
-
-        console.log("profileOptions: ", profileOptions);
-
         messageCont.innerHTML = `You are now successfully logged in. If the feed doesn't load automatically, click here: <a href="feed.html" title="Feed">Feed</a>`;
+        window.location.href = "feed.html";
 
-        // window.location.href = "feed.html";
-
-      } else {
-        messageCont.innerHTML = `<span class="error-message">Something went wrong. Please make sure your email and password are filled out correctly.</span>`;
       }
-    }
-    catch (error) {
+    } catch (error) {
       console.log(error);
+      messageCont.innerHTML = `<span class="error-message">${error}</span>`;
     }
   }
 });
