+++ js/tags/handleTags.mjs
@@ -1,268 +1,9 @@
-const usernameConst = localStorage.getItem("username");
-const token = localStorage.getItem("accessToken");
-
-const baseUrl = "https://api.noroff.dev/api/v1";
-const allPostsUrlEndPoint = "/social/posts?_author=true&_comments=true&_reactions=true";
-const allPostsUrl = baseUrl + allPostsUrlEndPoint;
-const createPostEndpoint = "/social/posts";
-
-const createPostUrl = baseUrl + createPostEndpoint;
-const followingPostsUrlEnd = "/social/posts/following";
-const followingPostsUrl = baseUrl + followingPostsUrlEnd;
-
-const inputTitle = document.querySelector("#post-title");
-const inputMessage = document.querySelector("#post-message");
-const submitButton = document.querySelector("#post-button");
-const form = document.querySelector(".publish-message-form");
-const feedbackCont = document.querySelector(".feedback-cont");
-const tagInput = document.querySelector("#tags");
-
-const feedContNew = document.querySelector(".feed-cont-new");
-const feedCont = document.querySelector(".feed-cont");
-
-form.addEventListener("submit", async (data) => {
-    data.preventDefault();
-
-    if (!inputTitle.value || !inputMessage.value) {
-        feedbackCont.innerHTML = `<span class="error-message">Please fill out your title and post message.</span>`;
-    }    
-
-    if (inputTitle.value || inputMessage.value) {
-        const titleConst = inputTitle.value;
-        const messageConst = inputMessage.value;
-        const tagConst = tagInput.value;
-        console.log("taginput: ", tagConst);
-
-        const createPostConst = {
-            method: 'POST',
-            body: JSON.stringify({
-              title: `${titleConst}`,
-              body: `${messageConst}`, 
-              tags: [`${tagConst}`]
-            }),
-            headers: {
-                "Content-type": 'application/json; charset=UTF-8',
-                "Authorization": `Bearer ${token}`
-            },
-        };
-
-        console.log("createPostConst: ", createPostConst);
-
-        try {
-            const resp = await fetch(createPostUrl, createPostConst);
-
-            console.log("resp const: ", resp);
-
-            const json = await resp.json();
-            console.log("json const: ", json);
-
-            if(resp.ok) {
-                feedbackCont.innerHTML = `<span class="success-message">Your post was successfully published.</span>`;
-            }
-
-            let postBody = json.body;
-            let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
-
-            let postTime = json.created;
-            let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
-
-            let postTitle = json.title;
-            console.log("title: ", postTitle);
-
-            let postId = json.id;
-            console.log("postID: ", postId);
-
-            let postTag = json.tags[0];
-            console.log("postTag: ", postTag);
-
-            const postCard = document.createElement("div");
-            postCard.classList.add("post-card");
-            postCard.innerHTML = `
-
-                <div class="post-card-header">
-                    <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
-                    </div>
-                    <div>
-                        <div class="name-poster-div">${usernameConst}</div>
-                        <div class="post-footer-time-cont">
-                            <div>Posted: ${postDate}</div>
-                        </div>
-                    </div>
-                    <div class="post-card-header-div">
-                        <div class="topic-tag-cont">
-                            <div class="topic-tag">${postTag}</div>
-                        </div>
-                    </div>
-                </div>
-                <div class="post-card-main">
-                    <h4 class="post-card-title"></h4>
-                    <div class="post-main-message"></div>
-                    <a href="post.html?postId=${postId}" class="see-full-button">
-                        <div class="see-full-div">See Full</div>
-                    </a>
-                </div>
-                <div class="post-card-footer-individual">
-                    <div class="post-footer-icon-cont">
-                        <img src="images/like-icon-heart.png" alt="Like icon" class="like-icon-class">
-                        <div>0 likes</div>
-                    </div>
-                    <div class="post-footer-icon-cont">
-                        <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
-                        <div>0 comments</div>
-                    </div>
-                </div>
-
-            `
-            postCard.querySelector(".post-card-title").textContent = postTitle;
-            postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
-            postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
-            feedContNew.appendChild(postCard);
-
-        }
-
-        catch (error) {
-            console.log("Error from catch error: ", error);
-        }
-    }  
-});
-
-const options = {
-    headers: {
-        Authorization: `Bearer ${token}`,
-    },
-};
-
-async function displayAllPosts() {
-    try {
-        const resp = await fetch(allPostsUrl, options);
-        console.log("Response for API call: ", resp);
-        const json = await resp.json();
-        console.log("JSON const: ", json);
-
-        console.log("JSON.length: ", json.length);
-
-        for(let i = 0; i < 20; i++) {
-
-            let postBody = json[i].body;
-            let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
-
-            let postAuthor = json[i].author.name;
-            console.log("postAuthor: ", postAuthor);
-
-            let postTime = json[i].created;
-            let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
-
-            let postTitle = json[i].title;
-            console.log("title: ", postTitle);
-
-            let postId = json[i].id;
-            console.log("postID: ", postId);
-
-            let postTag = json[i].tags[0];
-            console.log("postTag: ", postTag);
-            
-            const postCard = document.createElement("div")
-            postCard.classList.add("post-card")
-            postCard.innerHTML = `
-
-                <div class="post-card-header">
-                    <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
-                    </div>
-                    <div>
-                        <div class="name-poster-div"></div>
-                        <div class="post-footer-time-cont">
-                            <div>Posted: ${postDate}</div>
-                        </div>
-                    </div>
-                    <div class="post-card-header-div">
-                        <div class="topic-tag-cont">
-                            <div class="topic-tag"></div>
-                        </div>
-                    </div>
-                </div>
-                <div class="post-card-main">
-                    <h4 class="post-card-title"></h4>
-                    <div class="post-main-message"></div>
-                    <a href="post.html?postId=${postId}" class="see-full-button">
-                        <div class="see-full-div">See Full</div>
-                    </a>
-                </div>
-                <div class="post-card-footer-individual">
-                    <div class="post-footer-icon-cont">
-                        <img src="images/like-icon-heart.png" alt="Like icon" class="like-icon-class">
-                        <div>0 likes</div>
-                    </div>
-                    <div class="post-footer-icon-cont">
-                        <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
-                        <div>0 comments</div>
-                    </div>
-                </div>
-
-            `
-            postCard.querySelector(".post-card-title").textContent = postTitle;
-            postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
-            postCard.querySelector(".name-poster-div").textContent = postAuthor;
-            postCard.querySelector(".topic-tag").textContent = postTag;
-            postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
-            feedCont.appendChild(postCard);
-
-        }
-
-    }
-    catch (error) {
-        console.log("Console log error from catch error: ", error);
-    }
-};
-
-displayAllPosts();
-
-
-// Find post by ID code
-
-const findButton = document.querySelector("#find-button");
-const searchButton = document.querySelector("#search-button");
-const searchInput = document.querySelector("#search-input");
-const searchForm = document.querySelector(".search-container");
-
-
-findButton.addEventListener("click", findPostById);
-
-async function findPostById() {
-    event.preventDefault();
-    const idContainer = searchInput.value;
-
-    const apiSinglePostEndpoint = `/social/posts/${idContainer}posts?_author=true&_comments=true&_reactions=true`;
-    const apiUrlSinglePost = baseUrl + apiSinglePostEndpoint;
-
-    try {
-        const resp = await fetch(apiUrlSinglePost, options);
-        const json = await resp.json();
-        console.log("json from find by id: ", json);
-
-        if (resp.ok) {
-                window.location.href = `post.html?postId=${idContainer}`
-        } 
-        if (!resp.ok) {
-            findButton.style.borderColor = "red";
-            findButton.style.color = "red";
-            findButton.style.fontWeight = "bold";
-            searchInput.style.borderColor = "red";
-            searchInput.style.color = "red";
-            searchInput.style.fontWeight = "bold";
-        }
-
-    } catch(error) {
-        console.log("error from find by id: ", error);
-    }
-};
-
 // Filter posts by tags
+
+import { allPostsUrl, baseUrl,myPostsUrl } from "../constants/api.mjs";
+import { usernameConst, token, profilePictureUrl } from "../constants/localStorage.mjs";
+
+export { displayYourPosts, displayFilteredPostsProduct, displayFilteredPostsDiscovery, displayFilteredPostsDelivery, displayFilteredPostsTechnology, displayFilteredPostsInnovation, displayFilteredPostsFunding };
 
 const yourPostsTagCont = document.querySelector("#tag-div-your-posts");
 const productTagCont = document.querySelector("#tag-div-product");
@@ -271,29 +12,27 @@ const deliveryTagCont = document.querySelector("#tag-div-delivery");
 const technologyTagCont = document.querySelector("#tag-div-technology");
 const innovationTagCont = document.querySelector("#tag-div-innovation");
 const fundingTagCont = document.querySelector("#tag-div-funding");
+const feedCont = document.querySelector(".feed-cont");
 
-yourPostsTagCont.addEventListener("click", displayFilteredPostsYour);
-productTagCont.addEventListener("click", displayFilteredPostsProduct);
-discoveryTagCont.addEventListener("click", displayFilteredPostsDiscovery);
-deliveryTagCont.addEventListener("click", displayFilteredPostsDelivery);
-technologyTagCont.addEventListener("click", displayFilteredPostsTechnology);
-innovationTagCont.addEventListener("click", displayFilteredPostsInnovation);
-fundingTagCont.addEventListener("click", displayFilteredPostsFunding);
+const options = {
+    headers: {
+        Authorization: `Bearer ${token}`,
+    },
+};
 
-async function displayFilteredPostsYour() {
+// Handle "Your posts" tag filter
+// I didn't have time to rewrite these handleTags functions into one function to process the various filterings before the deadline.
+
+const displayYourPosts = yourPostsTagCont.addEventListener("click", async function handleTags() {
     try {
-        const resp = await fetch(allPostsUrl, options);
-        const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.author.name === usernameConst) {
-                return true;
-            }
+        const resp = await fetch(myPostsUrl, {
+            headers: {
+                "Content-type": "application/json; charset=UTF-8",
+                Authorization: `Bearer ${token}`,
+		    },
         });
-
-        console.log("filteredPosts: ", filteredPosts);
+        
+        const posts = await resp.json();
 
         if (resp.ok) {
             feedCont.innerHTML = "";
@@ -307,27 +46,27 @@ async function displayFilteredPostsYour() {
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
-
-            let postBody = filteredPosts[i].body;
+        for(let i = 0; i < posts.length; i++) {
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
@@ -335,7 +74,7 @@ async function displayFilteredPostsYour() {
 
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -365,38 +104,31 @@ async function displayFilteredPostsYour() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
-async function displayFilteredPostsProduct() {
+// Handle "Product" tag filter
+
+const displayFilteredPostsProduct = productTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Product") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
 
         if (resp.ok) {
             feedCont.innerHTML = "";
@@ -410,27 +142,27 @@ async function displayFilteredPostsProduct() {
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
-
-            let postBody = filteredPosts[i].body;
+        for(let i = 0; i < posts.length; i++) {
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
@@ -438,7 +170,7 @@ async function displayFilteredPostsProduct() {
 
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -468,40 +200,32 @@ async function displayFilteredPostsProduct() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
 
+// Handle "Product discovery" tag filter
 
-async function displayFilteredPostsDiscovery() {
+const displayFilteredPostsDiscovery = discoveryTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Product discovery") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
 
         if (resp.ok) {
             feedCont.innerHTML = "";
@@ -515,35 +239,35 @@ async function displayFilteredPostsDiscovery() {
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
+        for(let i = 0; i < posts.length; i++) {
 
-            let postBody = filteredPosts[i].body;
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
             postCard.innerHTML = `
-
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -573,81 +297,74 @@ async function displayFilteredPostsDiscovery() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
 
-async function displayFilteredPostsDelivery() {
+// Handle "Product delivery" tag filter
+
+const displayFilteredPostsDelivery = deliveryTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Product delivery") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
 
         if (resp.ok) {
             feedCont.innerHTML = "";
             deliveryTagCont.classList.add("marked-tag");
 
             yourPostsTagCont.classList.remove("marked-tag");
-            discoveryTagCont.classList.remove("marked-tag");
             productTagCont.classList.remove("marked-tag");
+            discoveryTagCont.classList.remove("marked-tag");
             technologyTagCont.classList.remove("marked-tag");
             innovationTagCont.classList.remove("marked-tag");
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
+        for(let i = 0; i < posts.length; i++) {
 
-            let postBody = filteredPosts[i].body;
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
             postCard.innerHTML = `
-
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -677,81 +394,75 @@ async function displayFilteredPostsDelivery() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
+// Handle "Technology" tag filter
 
-async function displayFilteredPostsTechnology() {
+const displayFilteredPostsTechnology = technologyTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
 
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Technology") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
+        console.log("posts: ", posts)
 
         if (resp.ok) {
             feedCont.innerHTML = "";
             technologyTagCont.classList.add("marked-tag");
 
             yourPostsTagCont.classList.remove("marked-tag");
+            productTagCont.classList.remove("marked-tag");
             discoveryTagCont.classList.remove("marked-tag");
             deliveryTagCont.classList.remove("marked-tag");
-            productTagCont.classList.remove("marked-tag");
             innovationTagCont.classList.remove("marked-tag");
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
+        for(let i = 0; i < posts.length; i++) {
 
-            let postBody = filteredPosts[i].body;
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
             postCard.innerHTML = `
-
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -781,81 +492,73 @@ async function displayFilteredPostsTechnology() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
+// Handle "Innovation" tag filter
 
-async function displayFilteredPostsInnovation() {
+const displayFilteredPostsInnovation = innovationTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Innovation") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
 
         if (resp.ok) {
             feedCont.innerHTML = "";
             innovationTagCont.classList.add("marked-tag");
 
             yourPostsTagCont.classList.remove("marked-tag");
+            productTagCont.classList.remove("marked-tag");
             discoveryTagCont.classList.remove("marked-tag");
             deliveryTagCont.classList.remove("marked-tag");
             technologyTagCont.classList.remove("marked-tag");
-            productTagCont.classList.remove("marked-tag");
             fundingTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
+        for(let i = 0; i < posts.length; i++) {
 
-            let postBody = filteredPosts[i].body;
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
             postCard.innerHTML = `
-
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -885,81 +588,73 @@ async function displayFilteredPostsInnovation() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
 
+// Handle "Innovation" tag filter
 
-async function displayFilteredPostsFunding() {
+const displayFilteredPostsFunding = fundingTagCont.addEventListener("click", async function handleTags(data) {
+    const myTag = data.target.dataset.tag;
+    const tagsUrl = `${baseUrl}/social/posts?_tag=${myTag}&_author=true&_comments=true&_reactions=true`;
+
     try {
-        const resp = await fetch(allPostsUrl, options);
+        const resp = await fetch(tagsUrl, options);
         const posts = await resp.json();
-
-        console.log("posts: ", posts);
-
-        const filteredPosts = posts.filter((post) => {
-            if(post.tags[0] === "Funding") {
-                return true;
-            }
-        });
-
-        console.log("filteredPosts: ", filteredPosts);
 
         if (resp.ok) {
             feedCont.innerHTML = "";
             fundingTagCont.classList.add("marked-tag");
 
             yourPostsTagCont.classList.remove("marked-tag");
+            productTagCont.classList.remove("marked-tag");
             discoveryTagCont.classList.remove("marked-tag");
             deliveryTagCont.classList.remove("marked-tag");
             technologyTagCont.classList.remove("marked-tag");
             innovationTagCont.classList.remove("marked-tag");
-            productTagCont.classList.remove("marked-tag");
         }
 
-        for(let i = 0; i < 20; i++) {
+        for(let i = 0; i < posts.length; i++) {
 
-            let postBody = filteredPosts[i].body;
+            let postBody = posts[i].body;
             let postBodyExcerpt = postBody.substring(0, 400);
-            console.log("postBodyExcerpt: ", postBodyExcerpt);
 
-            let postAuthor = filteredPosts[i].author.name;
-            console.log("postAuthor: ", postAuthor);
+            let postAuthor = posts[i].author.name;
 
-            let postTime = filteredPosts[i].created;
+            let postTime = posts[i].created;
             let postDate = postTime.substring(0, 10);
-            console.log("createdDate: ", postDate);
 
-            let postTitle = filteredPosts[i].title;
-            console.log("title: ", postTitle);
+            let postTitle = posts[i].title;
 
-            let postId = filteredPosts[i].id;
-            console.log("postID: ", postId);
+            let postId = posts[i].id;
 
-            let postTag = filteredPosts[i].tags[0];
-            console.log("postTag: ", postTag);
+            let postTag = posts[i].tags[0];
+
+            let postAuthorImage = posts[i].author.avatar;
+            if(postAuthorImage) {
+                postAuthorImage = posts[i].author.avatar;
+            } else {
+                postAuthorImage = "./images/blank-profile-picture.png";
+            }
             
             const postCard = document.createElement("div")
             postCard.classList.add("post-card")
             postCard.innerHTML = `
-
                 <div class="post-card-header">
                     <div class="post-card-header-div">
-                        <img src="" class="profile-picture-icon" alt="Profile pic">
+                        <img src="" class="profile-picture-icon-post" alt="Profile pic">
                     </div>
                     <div>
                         <div class="name-poster-div"></div>
@@ -989,20 +684,18 @@ async function displayFilteredPostsFunding() {
                         <img src="images/svg-comment.svg" alt="Comment icon" class="comment-icon-class">
                         <div>0 comments</div>
                     </div>
-                </div>
+                </div>`
 
-            `
             postCard.querySelector(".post-card-title").textContent = postTitle;
             postCard.querySelector(".post-main-message").textContent = postBodyExcerpt;
             postCard.querySelector(".name-poster-div").textContent = postAuthor;
             postCard.querySelector(".topic-tag").textContent = postTag;
             postCard.querySelector(".see-full-button").setAttribute("title", postTitle);
+            postCard.querySelector(".profile-picture-icon-post").setAttribute("src", postAuthorImage);
             feedCont.appendChild(postCard);
-
         }
-
     }
     catch (error) {
         console.log("Console log error from catch error: ", error);
     }
-};
+});
