+++ login.html
@@ -25,11 +25,11 @@
                     </div>
                     <div class="password-input-div">
                         <div class="password-feedback-container"></div>
-                        <input type="text" placeholder="Your password here*" id="password-input" name="password-input" class="password-input">
+                        <input type="password" placeholder="Your password here*" id="password-input" name="password-input" class="password-input">
                     </div>
-                    <div class=""> 
+                    <!-- <div class=""> 
                         <div class="forgot-pw-container"><a href="/X" title="Reset Password">Forgot password?</a></div>
-                    </div>
+                    </div> -->
                     <div class="button-div"> 
                         <button type="submit" class="submit-button-class" id="submit-button">Login</button>
                     </div>
@@ -41,6 +41,6 @@
         </main>
         <footer>
         </footer>
-        <script src="js/login.js"></script>
+        <script type="module" src="js/login.mjs"></script>
     </body>
 </html>
