+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "jsdoc": "^4.0.2"
+  }
+}
