+++ signup.html
@@ -33,11 +33,15 @@
                     </div>
                     <div class="">
                         <div class="password-feedback-container"></div>
-                        <input type="text" placeholder="Your password*" id="password-input" name="password" class="password-input">
+                        <input type="password" placeholder="Your password*" id="password-input" name="password" class="password-input">
                     </div>
-                    <div class="terms-accept-div">By signing up you consent to our <a href="/X" title="Terms and Conditions Privacy Policy">Terms</a> and <a href="/X" title="Privacy Policy Product Passion">Privacy Policy</a>
+                    <div class="">
+                        <div class="avatar-feedback-container"></div>
+                        <input type="text" placeholder="Optional: Profile picture live URL" id="avatar-input" name="avatar" class="avatar-input">
+                    </div>
+                    <!-- <div class="terms-accept-div">By signing up you consent to our <a href="/X" title="Terms and Conditions Privacy Policy">Terms</a> and <a href="/X" title="Privacy Policy Product Passion">Privacy Policy</a>
                         <input type="radio" id="consent-radio-button"></button>
-                    </div>
+                    </div> -->
                     <div class=""> 
                         <button type="submit" class="submit-button-class" id="submit-button">Sign Up</button>
                     </div>
@@ -49,6 +53,6 @@
         </main>
         <footer>
         </footer>
-        <script src="js/signup.js"></script>
+        <script type="module" src="js/signup.mjs"></script>
     </body>
 </html>
